+++ src/ts/Polygon.ts
@@ -220,7 +220,7 @@ export class Polygon implements SVGSerializable {
    * @return {number}
    */
   perimeter(): number {
-    let length = 0;
+    let length: number = 0;
     for (var i = 1; i < this.vertices.length; i++) {
       length += this.vertices[i - 1].distance(this.vertices[i]);
     }
@@ -272,47 +272,44 @@ export class Polygon implements SVGSerializable {
    */
   getEvenDistributionPolygon(pointCount: number): Polygon {
     if (pointCount <= 0) {
-      throw new Error("[getEvenDistributionPolygon] pointCount must be larger than zero; is " + pointCount + ".");
+      throw new Error(`pointCount must be larger than zero; is ${pointCount}.`);
     }
     var result = new Polygon([], this.isOpen);
     if (this.vertices.length === 0) {
       return result;
     }
-    var perimeter = this.perimeter();
-    var stepSize = perimeter / pointCount;
-    var segmentLength = 0;
 
     // Fetch and add the start point from the source polygon
     var polygonPoint = new Vertex(this.vertices[0]);
     result.vertices.push(polygonPoint);
-    var remainder = 0;
-    var polygonIndex = 1;
-    var loopMax = this.isOpen ? this.vertices.length : this.vertices.length + 1;
-    for (var i = 0; i < pointCount && polygonIndex < loopMax; i++) {
-      // Fetch next segment
-      var nextPolygonPoint = this.vertices[polygonIndex % this.vertices.length];
-      var segmentLength = polygonPoint.distance(nextPolygonPoint);
-      var segmentDiff = polygonPoint.difference(nextPolygonPoint);
-      var stepRatio = { x: segmentDiff.x / segmentLength, y: segmentDiff.y / segmentLength };
-      var j = 1;
-      while (segmentLength >= stepSize) {
-        // Get next point on local segment
-        var localStepSize = stepSize - (j === 1 ? remainder : 0);
-        var point = new Vertex(
-          polygonPoint.x + j * localStepSize * stepRatio.x - (j === 1 ? 0 : remainder) * stepRatio.x,
-          polygonPoint.y + j * localStepSize * stepRatio.y - (j === 1 ? 0 : remainder) * stepRatio.y
-        );
+    if (this.vertices.length === 1) {
+      return result;
+    }
 
-        // Don't add if we did a full turn here.
-        segmentLength -= localStepSize;
-        j++;
-        if (this.isOpen || polygonIndex + 1 < loopMax || segmentLength >= stepSize) {
-          result.vertices.push(point);
-        }
+    var perimeter = this.perimeter();
+    var stepSize = perimeter / (pointCount + 0);
+    var segmentLength = 0;
+    var n = this.vertices.length;
+
+    var polygonIndex = 1;
+    var nextPolygonPoint = new Vertex(this.vertices[1]);
+    var loopMax = this.isOpen ? n : n + 1;
+    var curSegmentU = stepSize;
+    var i = 1;
+    while (i < pointCount && polygonIndex < loopMax) {
+      var segmentLength = polygonPoint.distance(nextPolygonPoint);
+      // Check if next eq point is inside this segment
+      if (curSegmentU < segmentLength) {
+        var newPoint = polygonPoint.clone().lerpAbs(nextPolygonPoint, curSegmentU);
+        result.vertices.push(newPoint);
+        curSegmentU += stepSize;
+        i++;
+      } else {
+        polygonIndex++;
+        polygonPoint = nextPolygonPoint;
+        nextPolygonPoint = new Vertex(this.vertices[polygonIndex % n]);
+        curSegmentU = curSegmentU - segmentLength;
       }
-      polygonPoint = nextPolygonPoint;
-      remainder = segmentLength;
-      polygonIndex++;
     }
     return result;
   }